requirements.txt
requests==2.24.0
rpi.gpio==0.7.0
trio==0.16.0
